lambda-serving/arm_tvm/tvm/requirements.txt
numpy
decorator
attrs
tornado
xgboost
psutil
cloudpickle
boto3
pytest
python-dotenv
onnx
protobuf
Pillow
requests_toolbelt
